requirement.txt
amqp==2.1.1
Django==1.10.2
django-redis-cache==1.7.1
djangorestframework==3.5.1
hiredis==0.2.0
jsonfield==1.0.3
Markdown==2.6.7
pep8==1.7.0
python-dateutil==2.5.3
redis==2.10.5
six==1.10.0
vine==1.1.3
pika==0.10.0
